+++ atom/common/api/atom_api_native_image.cc
@@ -17,7 +17,7 @@
 #include "base/files/file_util.h"
 #include "base/strings/pattern.h"
 #include "base/strings/string_util.h"
-#include "chrome/browser/ui/webui/fileicon_source.h"
+#include "atom/common/fileicon_fetcher.h"
 #include "native_mate/dictionary.h"
 #include "native_mate/object_template_builder.h"
 #include "net/base/data_url.h"
@@ -306,7 +306,16 @@ gfx::Size NativeImage::GetSize() {
 void NativeImage::CreateFromFileIcon(v8::Isolate* isolate,
                                      const base::FilePath& path,
                                      const IconLoadedCallback& callback) {
-  callback.Run(CreateEmpty(isolate));
+  IconLoader::IconSize icon_size = IconLoader::IconSize::NORMAL;
+  float scale_factor = 1.0f;
+  auto onready = base::Bind(&NativeImage::OnIconLoaded, isolate, callback);
+  FileIconFetcher::FetchFileIcon(path, scale_factor, icon_size, onready);
+}
+
+void NativeImage::OnIconLoaded(v8::Isolate* isolate,
+                               const IconLoadedCallback& callback,
+                               gfx::Image& image) {
+  callback.Run(Create(isolate, image));
 }
 
 float NativeImage::GetAspectRatio() {
