+++ filenames.gypi
@@ -404,6 +404,8 @@
       'atom/common/crash_reporter/win/crash_service_main.h',
       'atom/common/draggable_region.cc',
       'atom/common/draggable_region.h',
+      'atom/common/fileicon_fetcher.cc',
+      'atom/common/fileicon_fetcher.h',
       'atom/common/google_api_key.h',
       'atom/common/key_weak_map.h',
       'atom/common/keyboard_util.cc',
@@ -547,8 +549,6 @@
       'chromium_src/chrome/browser/ui/views/color_chooser_aura.h',
       'chromium_src/chrome/browser/ui/views/frame/global_menu_bar_registrar_x11.cc',
       'chromium_src/chrome/browser/ui/views/frame/global_menu_bar_registrar_x11.h',
-      'chromium_src/chrome/browser/ui/webui/fileicon_source.cc',
-      'chromium_src/chrome/browser/ui/webui/fileicon_source.h',
       'chromium_src/chrome/common/chrome_constants.cc',
       'chromium_src/chrome/common/chrome_constants.h',
       'chromium_src/chrome/common/chrome_paths.cc',
